+++ src/errors/ProjectError.ts
@@ -1,3 +1,4 @@
+import { addEvent } from "../analytics";
 import { red } from "../utility/text";
 import { LoggableError } from "./LoggableError";
 
@@ -16,11 +17,13 @@ export enum ProjectErrorType {
 	BadRbxTypes,
 	BadRojoInclude,
 	TsMorph,
+	ProjectFailed,
 }
 
 export class ProjectError extends LoggableError {
 	constructor(message: string, public readonly type: ProjectErrorType) {
 		super(message);
+		void addEvent("ProjectError", ProjectErrorType[type]);
 	}
 
 	public log(projectPath: string) {